move method within extensions module itself
