+++ corehq/apps/registration/views.py
@@ -35,6 +35,7 @@ from corehq.apps.analytics.tasks import (
 from corehq.apps.analytics.utils import get_meta
 from corehq.apps.domain.decorators import login_required
 from corehq.apps.domain.exceptions import NameUnavailableException
+from corehq.apps.domain.extension_points import has_custom_clean_password
 from corehq.apps.domain.models import Domain
 from corehq.apps.hqwebapp.decorators import use_jquery_ui, use_ko_validation
 from corehq.apps.hqwebapp.views import BasePageView
@@ -51,7 +52,6 @@ from corehq.apps.registration.utils import (
     send_new_request_update_email,
 )
 from corehq.apps.users.models import CouchUser, WebUser
-from corehq.apps.users.views import has_custom_clean_password
 from corehq.const import USER_CHANGE_VIA_WEB
 from corehq.util.context_processors import get_per_domain_context
 from corehq.util.soft_assert import soft_assert
